Move join_terminate_close_proc to parsl.utils
This is used within parsl.monitoring.monitoring, but upcoming code
rearrangements to monitoring code want to use it in different places.

There is nothing monitoring specific about this - it should work with
any SpawnProcess at least (and probably any multiprocessing Process)
so this PR moves it to parsl.multiprocessing.
